# API/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.AgendamentoHojeResponse:
    properties:
      horario:
        example: "10:00"
        type: string
      id:
        example: 1
        type: integer
      observacoes:
        example: Paciente com alergia
        type: string
      paciente_nome:
        example: Maria Silva
        type: string
      procedimento_nome:
        example: Consulta Médica
        type: string
      status:
        example: Agendado
        type: string
      usuario_nome:
        example: Dr. João
        type: string
    type: object
  controllers.AgendamentoResumo:
    properties:
      data_hora:
        example: 2024-01-15 10:00
        type: string
      id:
        example: 1
        type: integer
      paciente:
        example: Maria Silva
        type: string
      procedimento:
        example: Consulta
        type: string
    type: object
...
  controllers.DashboardResponse:
    properties:
      agendamentos_hoje:
        example: 12
        type: integer
      agendamentos_semana:
        example: 45
        type: integer
      pacientes_ativos:
        example: 140
        type: integer
      proximos_agendamentos:
        items:
          $ref: '#/definitions/controllers.AgendamentoResumo'
        type: array
      receita_mes:
        example: 15000.5
        type: number
      total_pacientes:
        example: 150
        type: integer
      total_procedimentos:
        example: 25
        type: integer
      total_usuarios:
        example: 5
        type: integer
    type: object
...
  controllers.ErrorResponse:
    properties:
      error:
        example: Erro na operação
        type: string
    type: object
...
  controllers.EstatisticaMensal:
    properties:
      mes:
        example: 2024-01
        type: string
      total:
        example: 25
        type: integer
      valor:
        example: 5000
        type: number
    type: object
  controllers.EstatisticasResponse:
    properties:
      agendamentos_por_mes:
        items:
          $ref: '#/definitions/controllers.EstatisticaMensal'
        type: array
      pacientes_por_mes:
        items:
          $ref: '#/definitions/controllers.EstatisticaMensal'
        type: array
      procedimentos_populares:
        items:
          $ref: '#/definitions/controllers.ProcedimentoPopular'
        type: array
      receita_por_mes:
        items:
          $ref: '#/definitions/controllers.EstatisticaMensal'
        type: array
    type: object
...
  controllers.MetricasOperacionaisSwagger:
    properties:
      agendamentos_considerados:
        example: 120
        type: integer
      faturamento:
        example: 12500.5
        type: number
      periodo_fim:
        example: "2026-05-01"
        type: string
      periodo_inicio:
        example: "2026-04-01"
        type: string
      taxa_no_show_percentual:
        example: 6.67
        type: number
      total_faltas:
        example: 8
        type: integer
    type: object
...
  controllers.ProcedimentoPopular:
    properties:
      nome:
        example: Consulta Médica
        type: string
      total:
        example: 45
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API multiclínica com isolamento por tenant (clinic_id), RBAC por papel
    (ADM_GERAL, DONO, MEDICO, SECRETARIA), agendamento com validação de conflitos,
    prontuário com imutabilidade após 24h, auditoria LGPD e métricas de faturamento
    e no-show.
  title: API White Label — Clínicas
  version: "1.0"
paths:
...
  /convenios:
    post:
      consumes:
      - application/json
      description: Cadastrar um novo convênio
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cadastrar Convênio
      tags:
      - Convênios
  /convenios/{id}:
    put:
      consumes:
      - application/json
      description: Atualizar um convênio existente
      parameters:
      - description: ID do convênio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar Convênio
      tags:
      - Convênios
  /dashboard:
    get:
      consumes:
      - application/json
      description: Retorna métricas gerais da clínica
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.DashboardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Dashboard principal
      tags:
      - Dashboard
  /dashboard/agendamentos-hoje:
    get:
      consumes:
      - application/json
      description: Retorna os agendamentos do dia atual
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.AgendamentoHojeResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Agendamentos de hoje
      tags:
      - Dashboard
  /dashboard/estatisticas:
    get:
      consumes:
      - application/json
      description: Retorna estatísticas básicas da clínica
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.EstatisticasResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Estatísticas da clínica
      tags:
      - Dashboard
  /dashboard/metricas-operacionais:
    get:
      description: Faturamento pela soma do valor dos procedimentos agendados no período
        e taxa de absenteísmo (status Falta sobre agendamentos não cancelados).
      parameters:
      - default: primeiro dia do mês
        description: Data inicial (YYYY-MM-DD)
        in: query
        name: inicio
        type: string
      - description: Último dia inclusivo do período (YYYY-MM-DD)
        in: query
        name: fim
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MetricasOperacionaisSwagger'
      security:
      - BearerAuth: []
      summary: Métricas operacionais (faturamento e no-show)
      tags:
      - Dashboard
  /financeiro/abrir:
    get:
      consumes:
      - application/json
      description: Abrir módulo financeiro
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Abrir Financeiro
      tags:
      - Financeiro
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"